+++ src/include.spec.js
@@ -1,10 +1,22 @@
 const fs = require('fs-extra');
 const plugin = require('./include');
-const shared = require('./shared');
+const shared = require('./utils/shared');
+const jsonDataUtil = require('./utils/json-data');
 
 describe('Include Plugin', () => {
   beforeAll(() => {
     spyOn(shared, 'resolveAssetsPath').and.returnValue('');
+    spyOn(jsonDataUtil, 'parseAngularBindings').and.callFake(binding => {
+      if (binding.indexOf('mockFileOne') > -1) {
+        return 'test1.html'
+      }
+
+      if (binding.indexOf('mockFileTwo') > -1) {
+        return 'test2.html'
+      }
+
+      return binding;
+    });
   });
 
   it('should contain a preload hook', () => {
@@ -18,7 +30,7 @@ describe('Include Plugin', () => {
     expect(result.toString()).toEqual(content.toString());
   });
 
-  it('should not alter the content if the html file does not include any <stache include> tags.', () => {
+  it('should not alter the content if the html file does not include any <stache-include> tags.', () => {
     const content = new Buffer('<p></p>');
     const resourcePath = 'foo.html';
     const result = plugin.preload(content, resourcePath);
@@ -32,6 +44,28 @@ describe('Include Plugin', () => {
     const resourcePath = 'foo.html';
     const result = plugin.preload(content, resourcePath);
     expect(result.toString()).toContain(includeContents);
+  });
+
+  it('should not convert stache-include tags inside of a stache-code-block.', () => {
+    const includeContents = '<h1>Test</h1>';
+    spyOn(fs, 'readFileSync').and.returnValue(includeContents);
+    const content = new Buffer(`
+      <stache-code-block>
+        <stache-include fileName="test.html"></stache-include>
+      </stache-code-block>
+    `);
+    const resourcePath = 'foo.html';
+    const result = plugin.preload(content, resourcePath);
+    expect(result.toString()).not.toContain(includeContents);
+    expect(result.toString()).toEqual(content.toString());
+  });
+
+  it('It should replace stache.jsonData bindings in the fileName with the data value', () => {
+    const content = new Buffer('<stache-include fileName="{{ stache.jsonData.mock_data.mockFileOne }}"></stache-include>');
+    const includeContents = '<h1>Test1</h1>';
+    spyOn(fs, 'readFileSync').and.returnValue(includeContents);
+    const result = plugin.preload(content, 'foo.html');
+    expect(result.toString()).toEqual('<stache-include fileName="test1.html"><h1>Test1</h1></stache-include>');
   });
 
   it('should throw an error if the file is not found.', () => {
@@ -70,5 +104,29 @@ describe('Include Plugin', () => {
     const result = plugin.preload(content, resourcePath);
     expect(result.toString()).toContain(includeContents1);
     expect(result.toString()).toContain(includeContents2);
-  })
+  });
+
+  it('should support nested includes that use stache.jsonData bindings as fileNames', () => {
+    const includeComponent = '<stache-include fileName="{{ stache.jsonData.mock_data.mockFileTwo }}"></stache-include>';
+    const includeContents1 = '<h1>Test1</h1>';
+    const includeContents2 = '<h1>Test2</h1>';
+
+    spyOn(fs, 'readFileSync').and.callFake(file => {
+      if (file.indexOf('test1.html') > -1) {
+        return includeContents1 + includeComponent;
+      }
+
+      if (file.indexOf('test2.html') > -1) {
+        return includeContents2;
+      }
+
+      return '';
+    });
+
+    const content = new Buffer(`<stache-include fileName="{{ stache.jsonData.mock_data.mockFileOne }}"></stache-include>`);
+    const resourcePath = 'foo.html';
+    const result = plugin.preload(content, resourcePath);
+    expect(result.toString()).toContain(includeContents1);
+    expect(result.toString()).toContain(includeContents2);
+  });
 });