+++ src/json-data.js
@@ -1,51 +1,24 @@
-const fs = require('fs-extra');
-const path = require('path');
-const reserved = require('reserved-words');
-const shared = require('./shared');
-const glob = require('glob');
+const jsonDataUtil = require('./utils/json-data');
+const shared = require('./utils/shared');
 
 const preload = (content, resourcePath) => {
   if (resourcePath.match(/\.html$/)) {
-    return content.toString().replace(/stache\.jsonData\./g, 'stache.jsonData?.');
+    return addElvisOperator(content);
   }
 
-  if (!resourcePath.match(/app-extras\.module\.ts$/)) {
-    return content;
+  if (resourcePath.match(/app-extras\.module\.ts$/)) {
+    return addGlobalDataToAppExtrasModule(content, resourcePath);
   }
 
-  const root = shared.resolveAssetsPath('data');
-  const filePaths = glob.sync(path.join(root, '*.json'));
+  return content;
+};
 
-  if (!filePaths.length) {
-    return content;
-  }
+const addElvisOperator = (content) => {
+  return content.toString().replace(/\{\{\s*stache.jsonData./g, '{{ stache.jsonData?.');
+};
 
-  const dataObject = filePaths.reduce((acc, filePath) => {
-    const fileName = path.basename(filePath);
-    const propertyName = convertFileNameToObjectPropertyName(fileName);
-
-    if (!isPropertyNameValid(propertyName)) {
-      console.error(
-        new shared.StachePluginError(
-          `A valid Object property could not be determined from file ${fileName}! The property key '${propertyName}' cannot be used. Please choose another file name.`
-        )
-      );
-      return acc;
-    }
-
-    let contents;
-
-    try {
-      contents = fs.readFileSync(filePath);
-    } catch (error) {
-      throw new shared.StachePluginError(error.message);
-    }
-
-    acc[propertyName] = JSON.parse(contents);
-
-    return acc;
-  }, { });
-
+const addGlobalDataToAppExtrasModule = (content, resourcePath) => {
+  const globalData = jsonDataUtil.getGlobalData();
   const modulePath = shared.getModulePath(resourcePath);
 
   content = `
@@ -58,7 +31,7 @@ import {
 export const STACHE_JSON_DATA_PROVIDERS: any[] = [
   {
     provide: STACHE_JSON_DATA_SERVICE_CONFIG,
-    useValue: ${JSON.stringify(dataObject)}
+    useValue: ${JSON.stringify(globalData)}
   },
   {
     provide: StacheJsonDataService,
@@ -69,29 +42,6 @@ export const STACHE_JSON_DATA_PROVIDERS: any[] = [
 ${content}`;
 
   return shared.addToProviders(content, 'STACHE_JSON_DATA_PROVIDERS');
-};
-
-const convertFileNameToObjectPropertyName = (fileName) => {
-  return fileName.split('.')[0]
-    .replace(/\s+/g, '_')     // Replace spaces with underscores
-    .replace(/[^\w-]+/g, '')  // Remove all non-word chars
-    .replace(/-/g, '_')       // Replace all dashes with underscores
-    .replace(/--+/g, '_')     // Replace multiple dashes with single underscore
-    .replace(/^-+/, '')       // Trim - from start of text
-    .replace(/-+$/, '');      // Trim - from end of text;
-};
-
-const isPropertyNameValid = (propertyName) => {
-  if (!propertyName) {
-    return false;
-  }
-
-  if (propertyName === 'prototype') {
-    return false;
-  }
-
-  // Parsing as boolean because reserved-words returns `undefined` for falsy values.
-  return !reserved.check(propertyName, 'es6', true);
 };
 
 module.exports = { preload };